% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hinirm_no.R
\name{hinirm_no}
\alias{hinirm_no}
\title{title}
\usage{
hinirm_no(dataset, nschool, nmax, nitem, ndim = 2, ncount,
  niter = 30000, nburn = 5000, nthin = 5, nprint = 100,
  jump_beta = 0.1, jump_theta = 1, jump_z = 0.1, pr_mean_theta = 0,
  pr_sd_theta = 10, pr_mean_z = 0, pr_mean_gamma = 0,
  pr_sd_gamma = 10, pr_mean_mu = 0, pr_sd_mu = 10, prior_a = 0.001,
  prior_b = 0.001, option = TRUE, cores = 1)
}
\description{
title
}
