\name{nirm-package}
\alias{nirm-package}
\docType{package}
\title{Network Inductive Reasoning Model}
\description{
Fits Vardavas's Inductive Reasoning Model to a Network.
}
\details{
\tabular{ll}{
Package: \tab nirm\cr
Type: \tab Package\cr
Version: \tab 0.6.4a\cr
Date: \tab 2013-05-01\cr
License: \tab GPL 2.0 or greater\cr
LazyLoad: \tab yes\cr
}
Use this package to fit a an inductive reasoning model of a binary decision process to a network. 
}
\author{
Author: Raffaele Vardavas and Christopher Steven Marcum <cmarcum@uci.edu>
}
\keyword{ package }
