\name{nirm}
\alias{nirm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Network Inductive Reasoning Model.}
\description{Generate numeric categories of events in an idXevent list.}
\usage{
nirm(network.model=c("Erdos","Watts","NW.Watts","Barabasi","Empirical"),
               N=10^3,M=20,behavioral.model="Simple_Personal",enet=NULL,usbm=NULL,
               hyper.params=list(model=c(R0=2,gamma=1/3,iip=.02,eff=1,tfs=300,s.mem=.7),
               vattr=cbind(V=rep(0,N),w=rep(0.3,N),inf.nn=rep(0,N),vacc.nn=rep(0,N))),
               added.RMCN=FALSE, RMCN.R0=0,output.dir="",save.model=TRUE,verbose=FALSE,seed=NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{network.model}{A character string indicating the network generation process. Default is "Erdos." See Details.}
  \item{N}{An integer for the size of the network draw. If enet is non-null, N is taken from enet. Default is 10^3.}
  \item{M}{An integer for the average number of edges per vertice. Default is 20.}
  \item{behavioral.model}{A character string indicating the behavioral component of the model. Default is "Simple_Personal." See Details.}
  \item{enet}{If network.model="Empirical", an igraph object representing undirected relational network.}
  \item{usbm}{Optionally, a list containing the User Specified Behavioral Model. See details.}
  \item{hyper.params}{A list of parameters for the behavioral and network components of the model. See Details.}
  \item{added.RMCN}{Logical. Should we add a random-mixing component to the network model? Defaults to FALSE. See Details.}
  \item{RMCN.R0}{Numeric. If added.RMCN is TRUE, set the R0 for that network. Defaults to 0 (which is not used). Optional.}
  \item{output.dir}{A character string indicating where to save temporary and (optionally) permanent files. When non-null, this should contain a trailing "/". Default is pwd.}
  \item{save.model}{Logical. Should we save the final output to an R binary in the output.dir? Defaults to TRUE.}
  \item{seed}{Optionally, a random seed for the stochastic component of the model}
}

\details{
The \code{network.model} can be any of "Erdos","Watts","NW.Watts","Barabasi", or "Empirical." Each option corresponds to a specific network generation model. 
[[MORE DETAILS ABOUT EACH NETWORK MODEL HERE]]. The user may supply his or her own network data by setting this option to "Empirical" and supplying the appropriate
\pkg{igraph} object in \code{enet}.

The \code{behavioral.model} option specifies how actors in the network weight personal characteristics, information about the world, and the influence of their peers in their behaviors. Currently, internally constructed behavioral models are: "Original_Personal", "Simple_Personal", and "Simple_Local." [[EXPLAIN BEHAVIORAL MODEL HERE]].
Users can supply their own behavioral model by supplying a list of the proper structure to \code{usbm}. [[DESCRIBE USBM STRUCTURE HERE]].

Random mixing networks can be placed atop the network generated by the process by setting the \code{added.RMCN} to \code{TRUE}. The resulting network will always be a Bernoulli graph with O(N) and Pr(ij)=0.5. 
}

\value{
A list containing:
  \item{Results}{A list of seasons each containing an NX5 matrix, the columns of which are the results of the model. 
                 This length of this list is set by the \code{tfs} parameter in the \code{hyper.params}.}
  \item{Params}{A character string of the intitial parameters of the model.}
  \item{Runtime}{How long the model ran (in seconds).}
}
\note{Failing to provide a trailing slash ("/" or "\", depending on your platform) when specifying \code{output.dir} will result in by your temporary files possibly not being all that temporary :) }
\author{Raffaele Vardavas and Christopher Steven Marcum}
