% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnirme.R
\name{rnirme}
\alias{rnirme}
\title{title}
\usage{
rnirme(dataset, nsample, nitem, ndim = 2, niter = 30000,
  nburn = 5000, nthin = 5, nprint = 100, jump_beta = 0.2,
  jump_theta = 1, jump_w = 0.2, pr_mean_beta = 0, pr_sd_beta = 10,
  pr_mean_theta = 0, pr_sd_theta = 10, pr_mean_w = 0,
  prior_a = 0.001, prior_b = 0.001, option = TRUE, cores = 1)
}
\description{
title
}
