\name{nirm-package}
\alias{nirm-package}
\alias{nirm}
\docType{package}
\title{
\packageTitle{nirm}
}
\description{
\packageDescription{nirm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{nirm}
\packageIndices{nirm}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{nirm}

Maintainer: \packageMaintainer{nirm}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
